[REFACTOR/#48] #48 브랜치 관련 Gemini 코드 리뷰 적용
